[#20] chat update

+++ pages/chat/index.js
@@ -16,6 +16,7 @@ import { useRecoilValue } from "recoil";
 import { userState } from "../../state/atom/authState";
 import { DeleteRounded } from "@mui/icons-material";
 import { confirmAlert } from "../../components/common/Alert";
+import axios from "axios";
 
 const style = css`
   #chatlist {
@@ -95,16 +96,17 @@ export default function ChatList() {
   const userData = useRecoilValue(userState);
   let eventSource = useRef(null);
   const [listening, setListening] = useState(false);
+  const [deleteState, setDeleteState] = useState(true);
   const goChat = (id) => {
     router.push(`/chat/${id}`);
   };
-  const getChats = async () => {
+  const getChats = async (cancelToken) => {
     // setChatLists([]);
-
     try {
-      const getChatLists = await Api.get(
+      const getChatLists = await Api.getCancelToken(
         `http://20.214.170.222:8000/web-flux-service/api/v1/rooms/users/`,
-        userData.id
+        userData.id,
+        cancelToken
       );
 
       if (!getChatLists) {
@@ -118,9 +120,10 @@ export default function ChatList() {
           findUserInfos.push(user.userId);
         });
       });
-      const getUserInfos = await Api.get(
+      const getUserInfos = await Api.getCancelToken(
         `http://20.214.170.222:8000/user-service/api/v1/users/particular?param=`,
-        findUserInfos
+        findUserInfos,
+        cancelToken
       );
 
       const chats2 = getChatLists.data;
@@ -194,7 +197,7 @@ export default function ChatList() {
     }
   };
 
-  const getSse = async () => {
+  const getSse = async (cancelToken) => {
     if (content.current === undefined) {
       return;
     }
@@ -237,9 +240,10 @@ export default function ChatList() {
       if (content.current != undefined) {
         const copiedItems = [...chatLists];
         try {
-          const getChats = await Api.get(
+          const getChats = await Api.getCancelToken(
             `http://20.214.170.222:8000/web-flux-service/api/v1/rooms/`,
-            content.current.id
+            content.current.id,
+            cancelToken
           );
 
           if (!getChats) {
@@ -249,9 +253,10 @@ export default function ChatList() {
           if (Number(getChats.data.houseId) === 0) {
             getChats.data.roomInfo.users.forEach(async (user) => {
               if (Number(user.userId) !== Number(getSessionId())) {
-                const getUserInfos = await Api.get(
+                const getUserInfos = await Api.getCancelToken(
                   `http://20.214.170.222:8000/user-service/api/v1/users/particular?param=`,
-                  user.userId
+                  user.userId,
+                  cancelToken
                 );
                 if (!getUserInfos) {
                   throw new Error(`${user.userId} not allowd`);
@@ -286,8 +291,9 @@ export default function ChatList() {
   };
 
   useEffect(() => {
+    const cancleToken = axios.CancelToken.source();
     async function reactive() {
-      await getChats();
+      await getChats(cancleToken);
     }
     reactive();
     if (!listening) {
@@ -325,22 +331,28 @@ export default function ChatList() {
     return () => {
       eventSource.current.close();
       console.log("eventsource closed");
+      cancleToken.cancel();
     };
     // getChats();
     // connectSse();
   }, []);
   useEffect(() => {
+    const cancleToken = axios.CancelToken.source();
     // connectSse();
     // getChats();
     async function reactive() {
-      await getSse();
+      await getSse(cancleToken);
     }
     reactive();
     // getSse()
+
+    return () => {
+      cancleToken.cancel();
+    };
   }, [data]);
 
   const deletedEvent = (roomId, userId, username) => {
-    console.log(roomId + " :::" + userId);
+    if (deleteState) console.log(roomId + " :::" + userId);
     const sendAlert = {
       icon: "warning",
       title: username,
@@ -351,12 +363,20 @@ export default function ChatList() {
       .then(async (result) => {
         console.log(result);
         if (result) {
-          const chatDeleteResult = await Api.delete(
-            `http://20.214.170.222:8000/web-flux-service/api/v1/rooms/${userId}/users/${userId}`,
-            ""
-          );
-          console.log(chatDeleteResult);
-          setChatLists(chatLists.filter((item) => item.id !== roomId));
+          try {
+            if (deleteState) {
+              setDeleteState(false);
+              const chatDeleteResult = await Api.delete(
+                `http://20.214.170.222:8000/web-flux-service/api/v1/rooms/${roomId}/users/${userId}`,
+                ""
+              );
+              console.log(chatDeleteResult);
+              setChatLists(chatLists.filter((item) => item.id !== roomId));
+            }
+          } catch (e) {
+            console.log("error", e);
+          }
+          setDeleteState(true);
         }
       })
       .catch((res) => {
